requirements.txt
# Generated using pip freeze
PyYAML==6.0
urllib3==2.0.2